FIX: Optimization, project structure
